Add a reference cache context manager
Add a context manager for a reference cache so that it can be used
in a `with` statement. The resulting cache can be passed explicitly
to find_child or find_top_level, or it can be passed implicily because
the cache is stored on a hidden attribute on the document.