Work around a pySBOL3 bug
sbol3.Location does not use keyword arguments to call the parent constructor
and therefor cannot be part of a CustomIdentified. Use a type URI that is
in the SBOL3 namespace to work around this bug. The proper fix is a type
URI in the genbank namespace and inheriting from both sbol3.Location and
sbol3.CustomIdentified.